Avoid unnecessary reloading when called in dired-mode-hook
As functions in dired-mode-hook is run before dired reads in the file
list, it is more correct to use dired-sort-other with the second
argument (no-revert) set to t.

An extra function for this is added, and the formatting of switches
abstracted to a function to be called by both dired-quick-sort and the
new function dired-quick-sort-set-switches.